requirements.txt
aws-cdk.core==1.109.0
aws-cdk.aws-s3==1.109.0